LayoutTests/platform/blackberry/fast/forms/search-cancel-button-events-expected.txt

Test for event dipatching by search cancel button.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Initial state
PASS changeEventCounter is 0
PASS inputEventCounter is 0
Click the cancel button
FAIL search.value should be . Was foo.
PASS changeEventCounter is 0
FAIL inputEventCounter should be 1. Was 0.
Focus on another field
FAIL changeEventCounter should be 1. Was 0.
FAIL inputEventCounter should be 1. Was 0.
PASS successfullyParsed is true

TEST COMPLETE
